// src/charts/src/chart/model/chart-interface.ts
export interface IChartEventArgs {
    /** Defines the name of the event. */
    name: string;
    /** Defines the event's cancellation status. */
    cancel: boolean;
}
// ...
export interface IPointRenderEventArgs extends IChartEventArgs {
    /** Defines the series of the point. */
    series: Series;
    /** Defines the point. */
    point: Points;
    /** Defines the point fill color. */
    fill: string;
    /** Defines the point border. */
    border: BorderModel;
    /** Defines the point height. */
    height?: number;
    /** Defines the point width. */
    width?: number;
    /** Defines the point marker shape. */
    shape?: ChartShape;
    /** Defines the corner radius of the point. */
    cornerRadius?: CornerRadiusModel;
}